routes added to menu, student and course components created
