+++ src/frontend/src/courses/CourseDrawerForm.js
@@ -1,8 +1,8 @@
 import {Drawer, Input, Col, Select, Form, Row, Button, Spin} from 'antd';
-import {createStudent, getStudentById, updateStudent} from "./client";
+import {createStudent, getStudentById, updateStudent} from "../client";
 import {LoadingOutlined} from "@ant-design/icons";
 import {useEffect, useState} from "react";
-import {errorNotification, successNotification} from "./Notification";
+import {errorNotification, successNotification} from "../Notification";
 
 const {Option} = Select;
 const antIcon = <LoadingOutlined style={{ fontSize: 24 }} spin />;
